+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### ✨ Enhancements
 
+- Integrated Watchdog to enable automatic re-checking of Python files when changes are detected.
 - Added `autoformat-options` configuration option to let users specify command-line arguments to the Black formatting tool
 - Update `check_all` and `check_error` functions to let users pass in `typing.IO` objects to the `output` argument
 - Update the `forbidden-io-function-checker` to check functions from imported modules as well as methods (according to their qualified name)
@@ -19,6 +20,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### 🐛 Bug fixes
 
+- `check_contracts` no longer makes methods immediately enforce Representation Invariant checks when setting attributes of instances with the same type (one `Node` modifies another `Node` instance) and only checks RIs for these instances after the method returns.
 - Fixed error in `contracts` where comments in docstring assertions are not removed while parsing
 
 ### 🔧 Internal changes